refactor(wgpu): PMAT-357 split wgsl_forward into struct+layer files (Refs PMAT-357)
wgsl_forward.rs: 757→324 lines (struct, constructor, forward_model, upload)
wgsl_forward_layer.rs: 348 lines (forward_layer, encode helpers)
wgsl_shaders.rs: 127 lines (WGSL shader strings + BIAS_ADD_SHADER)

Fixed GPU LM head regression: upload_weight now stores biases in
weight_buffers too (for GPU bias_add), which triggered the GPU LM head
path with non-transposed weight. Reverted to CPU-only LM head.

60/60 contracts. Output verified correct: "2 + 2 equals 4."
